Further simplified API
