+++ api/controllers/index/logCallsRemaining.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const logCallsRemaining = (res) => {
+  const callsRemaining = res.meta['x-ratelimit-remaining'];
+
+  if (process.env.NODE_ENV !== 'production' || callsRemaining < 100) {
+    console.log(`API calls remaining: ${callsRemaining}`);
+  }
+
+  return res;
+};
+
+module.exports = logCallsRemaining;
