+++ api/controllers/index/prUtils.js
@@ -3,114 +3,6 @@
 const _ = require('lodash');
 const moment = require('moment');
 
-const prAmount = 5;
-
-const statements = [
-    'It\'s not too late to start!',
-    'Off to a great start, keep going!',
-    'Keep it up!',
-    'Nice! Now, don\'t stop!',
-    'So close!',
-    'Way to go!',
-    'Now you\'re just showing off!'
-];
-
-const errors = {
-    notUser: 'Username must belong to a user account.'
-};
-
-const errorCodes = {
-    notUser: 400
-};
-
-
-/**
- * GET /
- */
-exports.index = (req, res) => {
-    const github = req.app.get('github');
-    const username = req.query.username;
-
-    var hostname = process.env.APP_URL || `${req.protocol}://${req.headers.host}`;
-    var today = new Date();
-    var curmonth = today.getMonth();
-    var timeleft = 31 - today.getDate();
-    var timemessage = '';
-    if (curmonth === 9) {
-        if (timeleft === 0) {
-            timemessage = 'It\'s the very last day! Get your last PRs in!';
-        } else if (timeleft === 1) {
-            timemessage = 'One more day, keep it going!';
-        } else if (timeleft < 10) {
-            timemessage = 'There\'s only ' + timeleft + ' days left! You can do it!';
-        } else {
-            timemessage = 'There\'s ' + timeleft + ' days remaining!';
-        }
-    }
-
-    if (!username) {
-        if (req.xhr) {
-            return res.render('partials/error', {hostname: hostname, layout: false});
-        }
-
-        return res.render('index', {hostname: hostname, timemessage: timemessage});
-    }
-    function getStatement(prs) {
-        if (curmonth < 9) {
-            return 'Last year\'s result.';
-        } else if (curmonth === 9) {
-            return statements[prs.length < prAmount+1 ? prs.length : prAmount+1 ];
-        } else {
-            return 'This year\'s result.';
-        }
-    }
-
-    Promise.all([
-        findPrs(github, username),
-        github.users.getForUser({username})
-            .then(logCallsRemaining)
-    ])
-        .then(([prs, user]) => {
-            if (user.data.type !== 'User') {
-                return Promise.reject('notUser');
-            }
-
-            const data = {
-                prs,
-                isNotComplete: prs.length < prAmount,
-                statement: getStatement(prs),
-                username,
-                userImage: user.data.avatar_url,
-                hostname: hostname,
-                prAmount
-            };
-
-            if (req.query['plain-data']) {
-                res.render('partials/prs', _.assign(data, {layout: false}));
-            } else {
-                res.render('index', data);
-            }
-        }).catch((err) => {
-            console.log(err);
-            if (req.xhr) {
-                const code = errorCodes[err] || 404;
-                res.status(code).render('partials/error', {
-                    hostname: hostname,
-                    layout: false,
-                    errorMsg: errors[err]
-                });
-            } else {
-                res.render('index', {
-                    hostname: hostname,
-                    error: true,
-                    errorMsg: errors[err],
-                    username
-                });
-            }
-        });
-};
-
-
 let pullRequestData = [];
 
 function getNextPage(response, github) {
@@ -222,23 +114,3 @@ function findPrs(github, username) {
                 .then(mergeStatus => Promise.resolve(_.zipWith(prs, mergeStatus, (pr, merged) => _.assign(pr, {merged}))));
         });
 }
-
-const logCallsRemaining = res => {
-    var callsRemaining = res.meta['x-ratelimit-remaining'];
-    if (process.env.NODE_ENV !== 'production') {
-        console.log('API calls remaining: ' + callsRemaining);
-    } else if (callsRemaining < 100) {
-        console.log('API calls remaining: ' + callsRemaining);
-    }
-    return res;
-};
-
-exports.me = (req, res) => {
-    var hostname = process.env.APP_URL || `${req.protocol}://${req.headers.host}`;
-    res.render('me', {hostname: hostname});
-};
-
-exports.notfound = (req, res) => {
-    var hostname = process.env.APP_URL || `${req.protocol}://${req.headers.host}`;
-    res.render('404', {hostname: hostname});
-};
